do not use Cocoapods for building on iOS
I removed all the dependencies in the podfile so that everything is configured manually in the xcode project---exactly as it was before the upgrade.
Added @react-native-community/geolocation because it was removed from RN core and for some reason it's also a dependency in iOS (though I don't reallt know whjy it is required). Both Android and iOS should be correctly building at this point